Extract WordPressAuthenticator from `ViewController` in demo app
This is to keep the code a bit tidier.
