+++ Demo/AuthenticatorDemo/ViewController+WordPressAuthenticator.swift
@@ -1,0 +1,77 @@
+import WebKit
+import WordPressAuthenticator
+
+extension ViewController {
+
+    func initializeWordPressAuthenticator() {
+        // In a proper app, we'd want to split this call to keep the code readable. Here, it's
+        // useful to keep it all in one block to show how insanely long it is.
+        WordPressAuthenticator.initialize(
+            configuration: WordPressAuthenticatorConfiguration(
+                wpcomClientId: ApiCredentials.client,
+                wpcomSecret: ApiCredentials.secret,
+                wpcomScheme: "wordpress-authenticator-ios-demo",
+                wpcomTermsOfServiceURL: "https://wordpress.com/tos/",
+                wpcomBaseURL: "https://wordpress.com",
+                wpcomAPIBaseURL: "https://public-api.wordpress.com/",
+                googleLoginClientId: ApiCredentials.googleLoginClientId,
+                googleLoginServerClientId: ApiCredentials.googleLoginServerClientId,
+                googleLoginScheme: ApiCredentials.googleLoginSchemeId,
+                userAgent: "\(WKWebView.userAgent)-wordpress-authenticator-demo-app",
+                showLoginOptions: true,
+                enableSignUp: true,
+                // SIWA might require additional settings in the Developer Portal... Keeping it off
+                // for the moment
+                enableSignInWithApple: false,
+                enableSignupWithGoogle: true,
+                enableUnifiedAuth: true,
+                enableUnifiedCarousel: true
+            ),
+            style: WordPressAuthenticatorStyle(
+                // Primary (normal and highlight) is the color of buttons such as "Log in or signup
+                // with WordPress.com"
+                primaryNormalBackgroundColor: .orange,
+                primaryNormalBorderColor: .none,
+                primaryHighlightBackgroundColor: .brown,
+                primaryHighlightBorderColor: .none,
+                // Secondary (normal and highlight) is the color of buttons such as "Enter your
+                // existing site address" (the one just below "Log in or signup...") or "Continue
+                // with Google".
+                secondaryNormalBackgroundColor: .blue,
+                secondaryNormalBorderColor: .black,
+                secondaryHighlightBackgroundColor: .purple,
+                secondaryHighlightBorderColor: .black,
+                disabledBackgroundColor: .systemGray,
+                disabledBorderColor: .systemGray,
+                primaryTitleColor: .white,
+                secondaryTitleColor: .white,
+                disabledTitleColor: .white,
+                disabledButtonActivityIndicatorColor: .label,
+                textButtonColor: .red,
+                textButtonHighlightColor: .red,
+                instructionColor: .label,
+                subheadlineColor: .secondaryLabel,
+                placeholderColor: .red,
+                viewControllerBackgroundColor: .red,
+                textFieldBackgroundColor: .red,
+                // The navBar settings here are ignored. Those in
+                // `WordPressAuthenticatorUnifiedStyle` take precedence.
+                navBarImage: UIImage(),
+                navBarBadgeColor: .red,
+                navBarBackgroundColor: .orange
+            ),
+            unifiedStyle: WordPressAuthenticatorUnifiedStyle(
+                borderColor: .separator,
+                errorColor: .red,
+                textColor: .label,
+                textSubtleColor: .blue,
+                textButtonColor: .purple,
+                textButtonHighlightColor: .orange,
+                viewControllerBackgroundColor: .systemBackground,
+                navBarBackgroundColor: .blue,
+                navButtonTextColor: .white,
+                navTitleTextColor: .white
+            )
+        )
+    }
+}
